+++ src/mbgl/gfx/dynamic_texture_atlas.cpp
@@ -1,6 +1,7 @@
 #include <mbgl/gfx/dynamic_texture_atlas.hpp>
 #include <mbgl/gfx/context.hpp>
 
+#include <algorithm>
 #include <cmath>
 
 namespace mbgl {
@@ -11,16 +12,18 @@ constexpr const uint16_t padding = ImagePosition::padding + extraPadding;
 constexpr const Size startSize = {512, 512};
 constexpr const Size dummySize = {1, 1};
 
+namespace {
 Rect<uint16_t> rectWithoutExtraPadding(const Rect<uint16_t>& rect) {
     return Rect<uint16_t>(
         rect.x + extraPadding, rect.y + extraPadding, rect.w - 2 * extraPadding, rect.h - 2 * extraPadding);
 }
+} // namespace
 
 GlyphAtlas DynamicTextureAtlas::uploadGlyphs(const GlyphMap& glyphs,
                                              std::vector<std::function<void(Context&)>>& deletionQueue,
                                              const vk::UniqueCommandBuffer& commandBuffer) {
     using GlyphsToUpload = std::vector<std::tuple<TextureHandle, Immutable<Glyph>, FontStackHash>>;
-    std::lock_guard<std::mutex> lock(mutex);
+    std::scoped_lock lock(mutex);
 
     GlyphAtlas glyphAtlas;
     if (!glyphs.size()) {
@@ -99,8 +102,8 @@ GlyphAtlas DynamicTextureAtlas::uploadGlyphs(const GlyphMap& glyphs,
             glyphAtlas.dynamicTexture->uploadImage(paddedImage.data.get(), texHandle, deletionQueue, commandBuffer);
         }
         glyphAtlas.textureHandles.emplace_back(texHandle);
-        glyphAtlas.glyphPositions[fontStack].emplace(glyph->id,
-                                                     GlyphPosition{rectWithoutExtraPadding(rect), glyph->metrics});
+        glyphAtlas.glyphPositions[fontStack].emplace(
+            glyph->id, GlyphPosition{.rect = rectWithoutExtraPadding(rect), .metrics = glyph->metrics});
     }
     return glyphAtlas;
 }
@@ -111,7 +114,7 @@ ImageAtlas DynamicTextureAtlas::uploadIconsAndPatterns(const ImageMap& icons,
                                                        std::vector<std::function<void(Context&)>>& deletionQueue,
                                                        const vk::UniqueCommandBuffer& commandBuffer) {
     using ImagesToUpload = std::vector<std::pair<TextureHandle, Immutable<style::Image::Impl>>>;
-    std::lock_guard<std::mutex> lock(mutex);
+    std::scoped_lock lock(mutex);
 
     ImageAtlas imageAtlas;
     if (!icons.size() && !patterns.size()) {
@@ -247,7 +250,7 @@ ImageAtlas DynamicTextureAtlas::uploadIconsAndPatterns(const ImageMap& icons,
 
 void DynamicTextureAtlas::removeTextures(const std::vector<TextureHandle>& textureHandles,
                                          const DynamicTexturePtr& dynamicTexture) {
-    std::lock_guard<std::mutex> lock(mutex);
+    std::scoped_lock lock(mutex);
     if (!dynamicTexture) {
         return;
     }
@@ -256,7 +259,7 @@ void DynamicTextureAtlas::removeTextures(const std::vector<TextureHandle>& textu
         dynamicTexture->removeTexture(texHandle);
     }
     if (dynamicTexture->isEmpty()) {
-        auto iterator = std::find(dynamicTextures.begin(), dynamicTextures.end(), dynamicTexture);
+        auto iterator = std::ranges::find(dynamicTextures, dynamicTexture);
         if (iterator != dynamicTextures.end()) {
             dynamicTextures.erase(iterator);
         }