+++ src/mbgl/renderer/tile_pyramid.cpp
@@ -110,12 +110,12 @@ void TilePyramid::update(const std::vector<Immutable<style::LayerProperties>>& l
     std::vector<OverscaledTileID> idealTiles;
     std::vector<OverscaledTileID> panTiles;
 
-    util::TileCoverParameters tileCoverParameters = {parameters.transformState,
-                                                     parameters.tileLodMinRadius,
-                                                     parameters.tileLodScale,
-                                                     parameters.tileLodPitchThreshold};
+    util::TileCoverParameters tileCoverParameters = {.transformState = parameters.transformState,
+                                                     .tileLodMinRadius = parameters.tileLodMinRadius,
+                                                     .tileLodScale = parameters.tileLodScale,
+                                                     .tileLodPitchThreshold = parameters.tileLodPitchThreshold};
 
-    if (overscaledZoom >= zoomRange.min) {
+    if (std::cmp_greater_equal(overscaledZoom, zoomRange.min)) {
         int32_t idealZoom = std::min<int32_t>(zoomRange.max, overscaledZoom);
 
         // Make sure we're not reparsing overzoomed raster tiles.
@@ -158,7 +158,7 @@ void TilePyramid::update(const std::vector<Immutable<style::LayerProperties>>& l
 
     auto retainTileFn = [&](Tile& tile, TileNecessity necessity) -> void {
         if (retain.emplace(tile.id).second) {
-            tile.setUpdateParameters({minimumUpdateInterval, isVolatile});
+            tile.setUpdateParameters({.minimumUpdateInterval = minimumUpdateInterval, .isVolatile = isVolatile});
             tile.setNecessity(necessity);
         }