+++ src/mbgl/tile/tile_cache.cpp
@@ -72,7 +72,7 @@ void TileCache::deferPendingReleases() {
 
     // Block destruction until the cleanup task is complete
     {
-        std::lock_guard counterLock{deferredSignalLock};
+        std::scoped_lock counterLock{deferredSignalLock};
         deferredDeletionsPending++;
     }
 
@@ -91,7 +91,7 @@ void TileCache::deferPendingReleases() {
         MLN_ZONE_VALUE(tile_.items.size());
         tile_.items.clear();
 
-        std::lock_guard<std::mutex> counterLock(deferredSignalLock);
+        std::scoped_lock counterLock(deferredSignalLock);
         deferredDeletionsPending--;
         deferredSignal.notify_all();
     }};