+++ src/mbgl/tile/vector_mlt_tile.cpp
@@ -16,6 +16,9 @@ VectorMLTTile::VectorMLTTile(const OverscaledTileID& id_,
     : VectorTile(id_, std::move(sourceID_), parameters_, tileset_, observer_) {}
 
 VectorMLTTile::~VectorMLTTile() {
+    // this needs to be explicitly deleted in the most-derived destructor
+    loader.reset();
+
     // Don't rely on `~TileLoader` to close, it's not safe to call there.
     // We're still calling a virtual method from a destructor, so any overrides will not be called.
     GeometryTile::cancel();